Development

+++ data/migrations/20190830201522_expenses.js
@@ -1,25 +1,25 @@
 exports.up = function(knex) {
-  return knex.schema.createTable("expenses", tbl => {
+  return knex.schema.createTable('expenses', tbl => {
     tbl
-      .integer("user_id")
-      .references("id")
-      .inTable("users")
-      .onDelete("CASCADE")
-      .onUpdate("CASCADE")
+      .integer('user_id')
+      .references('id')
+      .inTable('users')
+      .onDelete('CASCADE')
+      .onUpdate('CASCADE')
       .notNullable();
 
-    tbl.increments("id");
-    tbl.integer("amount").notNullable();
+    tbl.increments('id');
+    tbl.decimal('amount').notNullable();
     tbl
-      .string("category", 128)
+      .string('category', 128)
       .notNullable()
       .unique();
-    tbl.text("notes", 500);
-    tbl.string("date").notNullable();
-    tbl.boolean("paid").defaultTo("false");
+    tbl.text('notes', 500);
+    tbl.string('date').notNullable();
+    tbl.boolean('paid').defaultTo('false');
   });
 };
 
 exports.down = function(knex) {
-  return knex.schema.dropTableIfExists("expenses");
+  return knex.schema.dropTableIfExists('expenses');
 };
